remove ballot_new.py
