+++ skule_vote/backend/test_ballot_real_data.py
@@ -3,8 +3,6 @@ import json
 
 from backend.ballot import calculate_results
 
-# from backend.ballot_new import calculate_results  # Uncomment to test with the new file
-
 from backend.models import (
     Ballot,
     Candidate,
